Rework of the debugger interface to use HANDLEs directly on Win32, rather than file descriptors, because file descriptors on Win32 are local to the runtime which causes problems when multiple runtimes are in play.
This isn't a full fix, because Glib still needs a FD to create the channel it
uses to listen for debugger input. This could be resolved by using a separate
thread to deal with accepting responses from the target and enqueuing them to
the main thread when ready. This would require using a custom event source to
connect this to the main loop.